// samples/Dressca/dressca-frontend/consumer/src/generated/api-client/models/paged-list-of-catalog-item-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Consumer Web API
 * Dressca Consumer の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CatalogItemResponse } from './catalog-item-response';

/**
 * ページネーションした結果のリストを管理します。             
 * @export
 * @interface PagedListOfCatalogItemResponse
 */
export interface PagedListOfCatalogItemResponse {
    /**
     * 現在のページ番号を取得します。 最初のページは 1 です。             
     * @type {number}
     * @memberof PagedListOfCatalogItemResponse
     */
    'page'?: number;
    /**
     * ページの総数を取得します。             
     * @type {number}
     * @memberof PagedListOfCatalogItemResponse
     */
    'totalPages'?: number;
    /**
     * 1 ページあたりの件数を取得します。             
     * @type {number}
     * @memberof PagedListOfCatalogItemResponse
     */
    'pageSize'?: number;
    /**
     * アイテムの総数を取得します。             
     * @type {number}
     * @memberof PagedListOfCatalogItemResponse
     */
    'totalCount'?: number;
    /**
     * 前のページが存在するかどうか示す値を取得します。             
     * @type {boolean}
     * @memberof PagedListOfCatalogItemResponse
     */
    'hasPrevious'?: boolean;
    /**
     * 次のページが存在するかどうか示す値を取得します。             
     * @type {boolean}
     * @memberof PagedListOfCatalogItemResponse
     */
    'hasNext'?: boolean;
    /**
     * 検索結果のリストを取得します。             
     * @type {Array<CatalogItemResponse>}
     * @memberof PagedListOfCatalogItemResponse
     */
    'items'?: Array<CatalogItemResponse>;
}